feat: add sleep and think funcions; reorder files
