+++ philo/philo.h
@@ -6,7 +6,7 @@
 /*   By: pmarkaid <pmarkaid@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/16 12:01:07 by pmarkaid          #+#    #+#             */
-/*   Updated: 2024/09/18 09:26:21 by pmarkaid         ###   ########.fr       */
+/*   Updated: 2024/09/18 09:38:01 by pmarkaid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,6 +53,7 @@ typedef struct s_table
 /* init */
 uint64_t get_time(void);
 t_table *init_table(int philos, int t_die, int t_eat, int t_sleep, int n_meals);
+void *routine(void *arg);
 void handle_routine(t_table *table);
 
 #endif /* PHILO_H */
